Support unique_param with value of None
There are reasons why a value might be set to None
(example when an IP address is to be created in Global VRF).
This commit fixes a bug where a unique_params with 'None' as the value is
ignored when splitting.

A value of "__not_set__" was chosen as default to avoid similar problem
if a default of "False" was chosen on fields boolean fields.