Support explicitly adding IPs to Global VRF
Default behavior is to add an IP to the Global VRF if no vrf is defined
as part of an IP address data set.
However this can result in issues if an IP needs to be added to
"Global" and a VRF when trying to "get_or_create".
This allows data to explicitly set a "vrf: null" in initializer data.