Enhancements to support more robust initializer data options (#18)
* Support unique_param with value of None

There are reasons why a value might be set to None
(example when an IP address is to be created in Global VRF).
This commit fixes a bug where a unique_params with 'None' as the value is
ignored when splitting.

A value of "__not_set__" was chosen as default to avoid similar problem
if a default of "False" was chosen on fields boolean fields.

* Support VRF configuration for primary_ips on devices

Without VRF support for device primary IPs, initializer data that uses
the same IP address in mulitple VRFs results in errors locating a single
IP object during assignment.

* Support explicitly adding IPs to Global VRF

Default behavior is to add an IP to the Global VRF if no vrf is defined
as part of an IP address data set.
However this can result in issues if an IP needs to be added to
"Global" and a VRF when trying to "get_or_create".
This allows data to explicitly set a "vrf: null" in initializer data.

* Define clusters before devices

Moved the initialization of clusters BEFORE devices
to allow the adding of devices to clusters while initializing them.

* Fix adding scope to vlan_group

The vlan_groups initializer wasn't correctly adding scopes configured
because the scope_type wasn't being added as a param. Fixed
issue by adding scope_type back in as a param.